Initial SMB service clustering work
Add basic registry wrappers around idmap, smb, active directory, and ldap
plugins. Disable groupmap and passdb for local users when clustered.

Make common smb.cluster_check method and add check for gluster enabled. Switch
to include registry and fix various places where the datastore / registry
wrapper wasn't being properly initialized. Wait in smb.configure until ctdb is
healthy before proceeding with SMB configuration.

Some service configuration changes errata:
-Added basic validation for clustered SMB shares to prevent enabling
 ZFS-specific features that are not available on glusterfs.
-Require cluster_volname if clustering enabled.
-Globally disable kernel share modes if clustering enabled (per vfs_glusterfs
  manpage).
-Ensure vfs_glusterfs is added to VFS objects list for clustered shares and
 that vfs objects are properly ordered
-Create separate per-volume log files in /var/log/samba4.
-Default log level to "1" (minimum) if entry is not present in global config

Added cluster health check to SMB service start routine smbd will panic in
some circustances if it is unable to establish a global ctdb messaging context.
This may happen while trying to intiialize a loadparm context with a clustered
registry backend while ctdb is stopped. Avoid this and other possibly
complications by simply refusing to start the SMB service if cluster is unhealthy.

Expand our enumeration of badness in smb.conf parameters

Fail in registry calls with ENXIO if ctdb unhealthy.

+++ src/middlewared/middlewared/plugins/smb_/passdb.py
@@ -70,6 +70,11 @@ class SMBService(Service):
         Accounts that are 'locked' in the UI will have their corresponding passdb entry
         disabled.
         """
+        ha_mode = await self.middleware.call('smb.get_smb_ha_mode')
+        if ha_mode == "CLUSTERED":
+            self.logger.debug("passdb support not yet implemented in clustered TrueNAS.")
+            return
+
         if passdb_backend is None:
             passdb_backend = await self.middleware.call('smb.getparm',
                                                         'passdb backend',
@@ -145,6 +150,11 @@ class SMBService(Service):
 
     @private
     async def remove_passdb_user(self, username):
+        ha_mode = await self.middleware.call('smb.get_smb_ha_mode')
+        if ha_mode == "CLUSTERED":
+            self.logger.debug("passdb support not yet implemented in clustered TrueNAS.")
+            return
+
         deluser = await run([SMBCmd.PDBEDIT.value, '-d', '0', '-x', username], check=False)
         if deluser.returncode != 0:
             raise CallError(f'Failed to delete user [{username}]: {deluser.stderr.decode()}')
@@ -182,6 +192,11 @@ class SMBService(Service):
         Delete any entries in the passdb_tdb file that don't exist in the config file.
         This method may cause temporary service disruption for SMB.
         """
+        ha_mode = await self.middleware.call('smb.get_smb_ha_mode')
+        if ha_mode == "CLUSTERED":
+            self.logger.debug("passdb support not yet implemented in clustered TrueNAS.")
+            return
+
         passdb_backend = await self.middleware.call('smb.getparm',
                                                     'passdb backend',
                                                     'global')
